requirements.txt
PyYAML>=6.0
mypy>=1.17.0
black>=25.1.0
isort>=6.0.0
types-PyYAML>=6.0.12